refactor: move textinput folder to text-input to respect convention
