+++ src/components/index.ts
@@ -1,5 +1,5 @@
 export { default as DropdownMenu } from './dropdown/DropdownMenu';
 export { default as DropdownElement } from './dropdown/DropdownElement';
 export { default as Tooltip } from './tooltip/Tooltip';
-export { TextInput } from './textinput/TextInput';
+export { TextInput } from './text-input/TextInput';
 export * from './tooltip/types';
